PM-35058 renamed files and removed unwanted promise.reject
