+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/onboarding/welcome-modal-dialog.component.ts
@@ -13,6 +13,7 @@ import {
   DialogService,
   TypographyModule,
 } from "@bitwarden/components";
+import { LogService } from "@bitwarden/logging";
 import { I18nPipe } from "@bitwarden/ui-common";
 
 import { OnboardingService } from "./services/onboarding.service";
@@ -24,8 +25,8 @@ import { OnboardingService } from "./services/onboarding.service";
   templateUrl: "./welcome-modal-dialog.component.html",
 })
 export class WelcomeModalDialogComponent {
-  private dialogRef = inject(DialogRef<WelcomeModalDialogComponent>);
-  private onboardingService = inject(OnboardingService);
+  private readonly dialogRef = inject(DialogRef<WelcomeModalDialogComponent>);
+  private readonly onboardingService = inject(OnboardingService);
 
   protected async onStartTour() {
     // invoke the dialog here
@@ -46,10 +47,14 @@ export class WelcomeModalDialogComponent {
     dialogService: DialogService,
   ): Promise<DialogRef<unknown, WelcomeModalDialogComponent>> {
     return runInInjectionContext(injector, async () => {
+      const logger = inject(LogService);
       const onboardingService = inject(OnboardingService);
       const acknowledged = await onboardingService.isWelcomeDialogAcknowledged();
       if (acknowledged) {
-        return Promise.reject("Welcome dialog already acknowledged.");
+        logger.info(
+          "[Access Intelligence Onboarding] Welcome dialog already acknowledged, skipping dialog display.",
+        );
+        return;
       }
 
       const dialog = dialogService.open(WelcomeModalDialogComponent, {
