+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/risk-insights.component.ts
@@ -59,7 +59,7 @@ import { ApplicationsComponent } from "./all-applications/applications.component
 import { CriticalApplicationsComponent } from "./critical-applications/critical-applications.component";
 import { EmptyStateCardComponent } from "./empty-state-card.component";
 import { RiskInsightsTabType } from "./models/risk-insights.models";
-import { WelcomeModalDialogComponent } from "./onboarding/welcome-modal-dialog.components";
+import { WelcomeModalDialogComponent } from "./onboarding/welcome-modal-dialog.component";
 import { DevMenuComponent } from "./shared/dev-menu.component";
 import { PageLoadingComponent } from "./shared/page-loading.component";
 import { ReportLoadingComponent } from "./shared/report-loading.component";
@@ -409,13 +409,7 @@ export class RiskInsightsComponent implements OnInit, OnDestroy {
 
   protected async beginOnboardingTour(): Promise<void> {
     if (this.adoptionUxImprovementsEnabled) {
-      await WelcomeModalDialogComponent.showWelcomeDialog(this.injector, this.dialogService).catch(
-        () => {
-          this.logService.info(
-            "Welcome dialog did not render. It was already acknowledged or this feature is not available.",
-          );
-        },
-      );
+      await WelcomeModalDialogComponent.showWelcomeDialog(this.injector, this.dialogService);
     }
   }
 }
